refactor Tuya Authorisation
